Refactor example fetching code to allow data-fetching code to run outside Vite
