+++ src/pages/latest.astro
@@ -1,8 +1,8 @@
 ---
-import { getExamples } from '../data/examples.js';
+import { getCategorizedExamples } from '../data/examples.js';
 import Layout from '../layouts/Base.astro';
 
-const examples = await getExamples();
+const examples = await getCategorizedExamples();
 ---
 
 <Layout examples={examples} basePath="/latest" />
