+++ src/pages/latest/[tab].astro
@@ -1,6 +1,6 @@
 ---
 import type { GetStaticPaths } from 'astro';
-import { type ExampleGroup, getExamples } from '../../data/examples.js';
+import { type ExampleGroup, getCategorizedExamples } from '../../data/examples.js';
 import Layout from '../../layouts/Base.astro';
 
 export type Props = {
@@ -8,7 +8,7 @@ export type Props = {
 };
 
 export const getStaticPaths: GetStaticPaths = async () => {
-	const examples = await getExamples();
+	const examples = await getCategorizedExamples();
 	return [...examples.entries()].map(([tab]) => ({
 		params: { tab },
 		props: { examples },
