+++ src/utils/github.ts
@@ -18,10 +18,10 @@ export function githubRequest(url: string) {
 	const headers: Headers = new Headers({
 		Accept: 'application/vnd.github.v3+json',
 	});
-	if (typeof import.meta.env.GITHUB_TOKEN !== 'string') {
+	if (typeof import.meta.env?.GITHUB_TOKEN !== 'string') {
 		console.warn(
 			`GITHUB_TOKEN is ${typeof import.meta.env
-				.GITHUB_TOKEN}. You may run into rate-limiting issues.`,
+				?.GITHUB_TOKEN}. You may run into rate-limiting issues.`,
 		);
 	} else {
 		headers.set('Authorization', `token ${import.meta.env.GITHUB_TOKEN}`);
